watsonx-indianinsurance-master/requirements.txt
streamlit
python-dotenv
st_clickable_images
streamlit_extras
ibm-generative-ai